+++ src/app/components/ui/ukhsa/List/ListItemStatus.tsx
@@ -1,9 +1,7 @@
-import clsx from 'clsx'
 import Link from 'next/link'
 import { ReactNode } from 'react'
 
 import { HealthAlertStatus, HealthAlertTypes } from '@/api/models/Alerts'
-import { useTranslation } from '@/app/i18n'
 
 import { ColdHealthAlertAmberIcon } from '../Icons/ColdHealthAlertsAmber'
 import { ColdHealthAlertGreenIcon } from '../Icons/ColdHealthAlertsGreen'
@@ -87,35 +85,5 @@ export const ListItemStatusTimestamp = ({ children }: ListItemStatusProps) => {
     <span className="govuk-body-s govuk-!-margin-bottom-0 govuk-!-margin-top-1 text-[14px] sm:text-[16px]">
       {children}
     </span>
-  )
-}
-
-interface ListItemStatusTagProps {
-  level: HealthAlertStatus | 'No alerts'
-  region: string
-  type: HealthAlertTypes
-}
-
-export async function ListItemStatusTag({ level, region, type }: ListItemStatusTagProps) {
-  const { t } = await useTranslation('weatherHealthAlerts')
-
-  let ariaLabel = t('statusLabelInitial', { level, type, region })
-
-  if (level === 'Amber') ariaLabel = t('statusLabelAmber', { level, type, region })
-
-  if (level === 'No alerts') ariaLabel = t('statusLabelNone', { level, type, region })
-
-  return (
-    <div
-      className={clsx('govuk-!-margin-right-0 govuk-phase-banner__content__tag m-auto capitalize', {
-        'govuk-tag govuk-tag--green': level === 'Green',
-        'govuk-tag govuk-tag--yellow': level === 'Yellow',
-        'govuk-tag govuk-tag--orange': level === 'Amber',
-        'govuk-tag govuk-tag--red': level === 'Red',
-      })}
-      aria-label={ariaLabel}
-    >
-      {level}
-    </div>
   )
 }
